requirements.txt
# Core ML dependencies
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
transformers>=4.38.0
accelerate>=0.27.0
safetensors>=0.4.2
tokenizers>=0.15.2
google-generativeai>=0.3.2

# Image processing
Pillow>=10.2.0
numpy>=1.26.0
opencv-python>=4.9.0.80
easyocr>=1.7.1
paddleocr>=2.7.0.3
paddlepaddle>=2.5.2
paddle2onnx>=1.1.3
protobuf==3.20.3  # Fixed version for PaddleOCR compatibility

# Web and API
gradio>=4.19.2
requests>=2.31.0
uvicorn>=0.27.1
websockets>=10.0,<12.0
websocket-client>=1.8.0

# Utilities
pydantic>=2.6.1
PyYAML>=6.0.1
python-dotenv>=1.0.1

# Platform-specific
pywin32>=306; sys_platform == 'win32'
uvloop>=0.21.0; sys_platform != 'win32'
